Fix renamed "server-xxx" files
